Moved components into the right directory
