+++ src/components/cellState.tsx
