+++ src/components/game-of-life.tsx
